Split Alpenglow voting_service from Tower voting_service.
