+++ core/src/lib.rs
@@ -36,7 +36,6 @@ pub mod sigverifier;
 pub mod sigverify_stage;
 pub mod snapshot_packager_service;
 pub mod staked_nodes_updater_service;
-pub mod staked_validators_cache;
 pub mod stats_reporter_service;
 pub mod system_monitor_service;
 pub mod tpu;
