+++ core/src/tvu.rs
@@ -17,7 +17,7 @@ use {
         repair::repair_service::{OutstandingShredRepairs, RepairInfo, RepairServiceChannels},
         replay_stage::{ReplayReceivers, ReplaySenders, ReplayStage, ReplayStageConfig},
         shred_fetch_stage::{ShredFetchStage, SHRED_FETCH_CHANNEL_SIZE},
-        voting_service::{VotingService, VotingServiceOverride},
+        voting_service::VotingService,
         warm_quic_cache_service::WarmQuicCacheService,
         window_service::{WindowService, WindowServiceChannels},
     },
@@ -54,6 +54,7 @@ use {
         event::{VotorEventReceiver, VotorEventSender},
         vote_history::VoteHistory,
         vote_history_storage::VoteHistoryStorage,
+        voting_service::{VotingService as AlpenglowVotingService, VotingServiceOverride},
         votor::LeaderWindowNotifier,
     },
     solana_votor_messages::consensus_message::ConsensusMessage,
@@ -85,6 +86,7 @@ pub struct Tvu {
     blockstore_cleanup_service: Option<BlockstoreCleanupService>,
     cost_update_service: CostUpdateService,
     voting_service: VotingService,
+    alpenglow_voting_service: AlpenglowVotingService,
     warm_quic_cache_service: Option<WarmQuicCacheService>,
     drop_bank_service: DropBankService,
     duplicate_shred_listener: DuplicateShredListener,
@@ -387,12 +389,16 @@ impl Tvu {
 
         let voting_service = VotingService::new(
             voting_receiver,
-            bls_receiver,
             cluster_info.clone(),
             poh_recorder.clone(),
             tower_storage,
-            vote_history_storage.clone(),
             vote_connection_cache.clone(),
+        );
+
+        let alpenglow_voting_service = AlpenglowVotingService::new(
+            bls_receiver,
+            cluster_info.clone(),
+            vote_history_storage,
             alpenglow_connection_cache,
             bank_forks.clone(),
             voting_service_test_override,
@@ -446,6 +452,7 @@ impl Tvu {
             blockstore_cleanup_service,
             cost_update_service,
             voting_service,
+            alpenglow_voting_service,
             warm_quic_cache_service,
             drop_bank_service,
             duplicate_shred_listener,
@@ -466,6 +473,7 @@ impl Tvu {
         }
         self.cost_update_service.join()?;
         self.voting_service.join()?;
+        self.alpenglow_voting_service.join()?;
         if let Some(warmup_service) = self.warm_quic_cache_service {
             warmup_service.join()?;
         }
