+++ votor/src/lib.rs
@@ -10,9 +10,11 @@ mod consensus_pool_service;
 pub mod event;
 mod event_handler;
 pub mod root_utils;
+mod staked_validators_cache;
 mod timer_manager;
 pub mod vote_history;
 pub mod vote_history_storage;
+pub mod voting_service;
 pub mod voting_utils;
 pub mod votor;
